refactor: update create-article feature
